api: oci image pull: do not pull OCI image if file already exists
This ensures that the API call fails early, before pulling the OCI image
from the registry and then failing to rename the temporary file.

Link: https://lore.proxmox.com/20251117171528.262443-3-f.schauer@proxmox.com

+++ src/PVE/API2/Storage/Status.pm
@@ -964,6 +964,8 @@ __PACKAGE__->register_method({
         my $path = PVE::Storage::get_vztmpl_dir($cfg, $storage);
         PVE::Storage::activate_storage($cfg, $storage);
 
+        die "refusing to override existing file '$filename'\n" if (-f "$path/$filename");
+
         local $SIG{INT} = sub {
             unlink "$path/$tmp_filename"
                 or warn "could not cleanup temporary file: $!"
